root/requirements.txt
anyjson==0.3.3
Django==1.11.7
django-cors-headers==2.1.0
django-utils2==2.4.0
djangorestframework==3.7.3
python-utils==2.2.0
pytz==2017.3
six==1.11.0
